[Refactor] side filter bar 컴포넌트화 완료
